move class hotslogsprovider to own package
