+++ src/test/java/ninja/eivind/hotsreplayuploader/providers/HotSLogsProviderTest.java
@@ -2,6 +2,7 @@ package ninja.eivind.hotsreplayuploader.providers;
 
 import com.amazonaws.services.s3.AmazonS3Client;
 import ninja.eivind.hotsreplayuploader.models.ReplayFile;
+import ninja.eivind.hotsreplayuploader.providers.hotslogs.HotsLogsProvider;
 import ninja.eivind.hotsreplayuploader.utils.SimpleHttpClient;
 import ninja.eivind.stormparser.StormParser;
 import ninja.eivind.stormparser.models.Replay;
